rename operations -> drawingOperations
because adding `annotationOperation`
