+++ src/api/drawingOperations.ts
